+++ tests/testthat/test-run.R
@@ -261,17 +261,7 @@ test_that("run deterministic parameterises model correctly", {
     100000,
     1000000
   )
-  dur_R <- 2.09
-  dur_hosp <- 5
-  probs <- default_probs()
-  m <- process_contact_matrix_scaled_age(m, pop$n)
-  beta <- beta_est_explicit(dur_R, dur_hosp, probs$prob_hosp, m, 3)
-  m <- t(t(m) / pop$n)
-  mix_mat_set <- aperm(array(c(m), dim = c(dim(m), 1)), c(3, 1, 2))
   expect_equal(length(output$output[,1]), 365)
-  expect_equal(output$parameters$beta, c(beta, beta/2))
-  expect_equal(output$parameters$tt_beta, c(0, 50))
-  expect_equal(output$parameters$mix_mat_set, mix_mat_set)
   expect_equal(output$parameters$hosp_bed_capacity, 100000)
   expect_equal(output$parameters$ICU_bed_capacity, 1000000)
 })
@@ -282,5 +272,15 @@ test_that("run day_return", {
   tp <- 10
   r <- run_explicit_SEEIR_model("Angola",replicates = 1, day_return = TRUE, time_period = tp)
   expect_true(nrow(r$output) == tp+1)
+
+})
+
+test_that("default probs", {
+
+  probs <- default_probs()
+  expect_named(probs,
+               c("prob_hosp","prob_severe","prob_non_severe_death_treatment",
+                 "prob_non_severe_death_no_treatment","prob_severe_death_treatment",
+                 "prob_severe_death_no_treatment","p_dist"))
 
 })