+++ NEWS.md
@@ -1,3 +1,8 @@
+# squire 0.4.3
+
+* `calibrate` can now take `squire:::deterministic_model()` as model type.
+* Output of above does not work with plotting or projections yet. 
+
 # squire 0.4.2
 
 * Patch for `plot.scan_results` where log likelihood was being plotted for both